consolidated azure functions
